+++ pkgs/default.nix
@@ -2,10 +2,17 @@
 
 let
 
-  nixpkgs = builtins.fetchGit {
-    name = "nixpkgs-20.03-snapshot";
-    url = "git@github.com:nixos/nixpkgs.git";
-    rev = "3f690bfcd4adde6dd0733c2d9f8f4d61e09dfc60";
+  nixpkgs = import ./patch-nixpkgs.nix {
+    src = builtins.fetchGit {
+      name = "nixpkgs-20.09";
+      url = "git@github.com:nixos/nixpkgs.git";
+      rev = "cd63096d6d887d689543a0b97743d28995bc9bc3";
+      ref = "refs/tags/20.09";
+    };
+    patches = [
+      # Fixed on *master* but not on *nixos-20.09*, as of 2020/11/30
+      ./patches/fix-lvm2-warnings-on-activation.patch
+    ];
   };
 
   overlay =
@@ -37,17 +44,8 @@ let
       breeze-contrast-cursor-theme = super.callPackage ./breeze-contrast-cursor-theme {};
 
       ocamlPackages = super.ocamlPackages.overrideScope' (self: super: {
-
-        hmap = self.callPackage ./ocaml-modules/hmap {};
-
         semver = self.callPackage ./ocaml-modules/semver {};
-
-        opium_kernel = self.callPackage ./ocaml-modules/opium_kernel {};
-        opium = self.callPackage ./ocaml-modules/opium {};
-
         obus = self.callPackage ./ocaml-modules/obus {};
-
-        mustache = self.callPackage ./ocaml-modules/mustache {};
       });
 
       # Controller