+++ controller/Changelog.md
@@ -7,6 +7,7 @@
 ## Changed
 
 - controller: Format machine-id in groups of 4 for readability
+- os: Update nixpkgs channel to 20.09
 
 ## Fixed